+++ src/hooks/LevelEditorLayer.cpp
@@ -2,6 +2,7 @@
 #include "../group-labels/Include.hpp"
 #include "../shared/Cache.hpp"
 #include "../shared/Settings.hpp"
+#include "../utils/Utils.hpp"
 
 using namespace geode::prelude;
 
@@ -44,5 +45,18 @@ class $modify(LevelEditorLayer) {
         if (label && !Settings::triggerWhitelist.contains(trigger->m_objectID)) {
             label->setVisible(false);
         }
+    }
+
+    GameObject* createObject(int p0, CCPoint p1, bool p2) {
+        GameObject* ret = LevelEditorLayer::createObject(p0, p1, p2);
+        if (auto trigger = typeinfo_cast<EffectGameObject*>(ret)) {
+            if (
+                Utils::triggerBlacklist.contains(trigger->m_objectID) ||
+                Settings::triggerWhitelist.contains(trigger->m_objectID)
+            ) return ret;
+            
+            trigger->m_hasSpecialChild = true;
+        }
+        return ret;
     }
 };